requirements.txt
fastapi
uvicorn
python-multipart
PILLOW
ultralytics
requests
